Extract thread conversation path helpers
